Reviewed all proposed changes numerous times with caytchen and Ubiquitatis.  Only remaining issue is with include/eve-server/inventory/AttributeMgr.h line 101, which I will fix in Ubi's absence.

+++ include/eve-server/ServiceDB.h
@@ -45,7 +45,7 @@ struct AccountInfo
     std::string name;
     std::string password;
     std::string hash;
-    uint64 id;
+    uint32 id;
     uint64 role;
     uint32 visits;
     std::string last_login;
