+++ include/eve-server/system/SystemEntity.h
@@ -99,7 +99,7 @@ public:
     virtual const GVector &GetVelocity() const = 0;
 	//get other attributes of the entity:
 	virtual const char *GetName() const = 0;
-	virtual double GetRadius() const = 0;
+	virtual float GetRadius() const = 0;
 
 	//I am not sure if I want this here...
 	virtual InventoryItemRef Item() const = 0;
@@ -147,7 +147,7 @@ public:
 	virtual const char *GetName() const;
 	virtual const GPoint &GetPosition() const;
 	virtual const GVector &GetVelocity() const;
-	virtual double GetRadius() const;
+	virtual float GetRadius() const;
 	virtual PyDict *MakeSlimItem() const;
 	virtual void MakeDamageState(DoDestinyDamageState &into) const;
