+++ src/eve-server/inventory/InventoryDB.cpp
@@ -711,7 +711,7 @@ bool InventoryDB::LoadTypeAttributes(uint32 typeID, EVEAttributeMgr &into) {
     
     for (; itr != attrset->end(); itr++) {
         if ((*itr)->number.get_type() == evil_number_int)
-            into.SetInt((EVEAttributeMgr::Attr)(*itr)->attributeID, (*itr)->number.get_int());
+            into.SetInt((EVEAttributeMgr::Attr)(*itr)->attributeID, static_cast<int32>((*itr)->number.get_int()));
         else
             into.SetReal((EVEAttributeMgr::Attr)(*itr)->attributeID, (*itr)->number.get_float());
     }
@@ -1747,28 +1747,28 @@ bool InventoryDB::GetOpenPowerSlots(uint32 slotType, ShipRef ship, uint32 &into)
 		attributeID = 1137;
 		firstFlag = 92; //rigslot0
 		//slotsOnShip = ship->rigSlots();
-        slotsOnShip = ship->GetAttribute(AttrRigSlots).get_int();
+        slotsOnShip = static_cast<uint32>(ship->GetAttribute(AttrRigSlots).get_int());
 	}
 	else if( slotType == 1 )
 	{
 		attributeID = 12;
 		firstFlag = 11; //lowslot0
 		//slotsOnShip = ship->lowSlots();
-        slotsOnShip = ship->GetAttribute(AttrLowSlots).get_int();
+        slotsOnShip = static_cast<uint32>(ship->GetAttribute(AttrLowSlots).get_int());
 	}
 	else if( slotType == 2 )
 	{
 		attributeID = 13;
 		firstFlag = 19; //medslot0
 		//slotsOnShip = ship->medSlots();
-        slotsOnShip = ship->GetAttribute(AttrMedSlots).get_int();
+        slotsOnShip = static_cast<uint32>(ship->GetAttribute(AttrMedSlots).get_int());
 	}
 	else if( slotType == 3 )
 	{
 		attributeID = 14;
 		firstFlag = 27; //hislot0
 		//slotsOnShip = ship->hiSlots();
-        slotsOnShip = ship->GetAttribute(AttrHiSlots).get_int();
+        slotsOnShip = static_cast<uint32>(ship->GetAttribute(AttrHiSlots).get_int());
 	}	
 	
 	for( uint32 flag = firstFlag; flag < (firstFlag + slotsOnShip); flag++ )
